+++ packages/components/stepper/src/step-number.tsx
